report coingecko balances directly on dfs and newdex
