+++ projects/newdex/index.js
@@ -1,6 +1,4 @@
-const axios = require("axios");
-const retry = require('../helper/retry')
-const { get_currency_balance } = require("../helper/eos");
+const { get_account_tvl } = require("../helper/eos");
 
 const tokens = [
   ["eosio.token", "EOS", "eos"],
@@ -16,19 +14,12 @@ const tokens = [
   ["core.ogx", "OGX", "organix"],
 ]
 
-async function get_account_tvl(account) {
-  const balances = {}
-  for ( const [ code, symbol, id ] of tokens ) {
-    const balance = await get_currency_balance(code, account, symbol);
-    balances[id] = balance
-  }
-  return balances;
-}
+
 
 // https://newdex.io
 // NewDex limit orderbook
 async function eos() {
-  return await get_account_tvl("newdexpublic");
+  return await get_account_tvl("newdexpublic", tokens);
 }
 
 module.exports = {
